refactor + update populate
