+++ miqa/core/rest/session.py
@@ -1,93 +1,12 @@
-import os
-from pathlib import Path
-import re
-
 from drf_yasg.utils import no_body, swagger_auto_schema
-from jsonschema import validate
-from jsonschema.exceptions import ValidationError as JSONValidationError
 from rest_framework import serializers, status
 from rest_framework.decorators import action
-from rest_framework.exceptions import ValidationError
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
-from miqa.core.conversion.csv_to_json import csvContentToJsonObject
-from miqa.core.models import Experiment, Image, Scan, ScanNote, Session, Site
-from miqa.core.models.scan import ScanDecision
-from miqa.core.schema.data_import import schema
-
-
-def import_data(user, session):
-    with open(session.import_path) as fd:
-        csv_content = fd.read()
-        try:
-            json_content = csvContentToJsonObject(csv_content)
-            validate(json_content, schema)  # TODO this should be an internal error
-        except (JSONValidationError, Exception) as e:
-            raise ValidationError({'error': f'Invalid CSV file: {str(e)}'})
-
-    data_root = Path(json_content['data_root'])
-
-    sites = {
-        site['name']: Site.objects.get_or_create(name=site['name'], defaults={'creator': user})[0]
-        for site in json_content['sites']
-    }
-
-    Experiment.objects.filter(session=session).delete()  # cascades to scans -> images, scan_notes
-
-    experiments = {
-        e['id']: Experiment(name=e['id'], note=e['note'], session=session)
-        for e in json_content['experiments']
-    }
-    Experiment.objects.bulk_create(experiments.values())
-
-    scans = []
-    images = []
-    notes = []
-    for scan_json in json_content['scans']:
-        experiment = experiments[scan_json['experiment_id']]
-        site = sites[scan_json['site_id']]
-        scan = Scan(
-            scan_id=scan_json['id'],
-            scan_type=scan_json['type'],
-            decision=ScanDecision.from_rating(scan_json['decision']),
-            experiment=experiment,
-            site=site,
-        )
-        scans.append(scan)
-
-        # TODO import notes
-        # if scan_json['note']:
-        #     notes_json = json.loads(unquote(scan_json['note']))
-        #     for note_json in notes_json:
-        #         scan_note = ScanNote(
-        #             **note_json,
-        #             scan=scan,
-        #         )
-        #         # This forces the modified field to use the value we give it
-        #         scan_note.update_modified = False
-        #         notes.append(scan_note)
-
-        if 'images' in scan_json:
-            # TODO implement this
-            raise Exception('use image_pattern for now')
-        elif 'image_pattern' in scan_json:
-            image_pattern = re.compile(scan_json['image_pattern'])
-            image_dir = data_root / scan_json['path']
-            for image_file in os.listdir(image_dir):
-                if image_pattern.fullmatch(image_file):
-                    images.append(
-                        Image(
-                            name=image_file,
-                            raw_path=image_dir / image_file,
-                            scan=scan,
-                        )
-                    )
-
-    Scan.objects.bulk_create(scans)
-    Image.objects.bulk_create(images)
-    ScanNote.objects.bulk_create(notes)
+from miqa.core.models import Session
+from miqa.core.tasks import import_data
 
 
 class SessionSerializer(serializers.ModelSerializer):
